Some renames to make things clearer
