+++ src/main/scala/org/uqbar/thin/javaless/JavalessCompiler.scala
@@ -4,8 +4,8 @@ import org.uqbar.voodoo.writer.ClassWriter
 import org.uqbar.voodoo.model._
 import org.uqbar.voodoo.mutator._
 
-object Compiler extends Compiler
-trait Compiler {
+object JavalessCompiler extends JavalessCompiler
+trait JavalessCompiler {
 	def compile(program: Program,targetPath: String) = 
     for(classDefinition <- program.definitions)
 		ClassWriter.writeClass($(classDefinition.name) let {it => }, targetPath)
